refactor: mv some services to dialogs
