+++ packages/cloud-toolkit/test/services/action-runner.unit.test.ts
@@ -2,7 +2,7 @@ import { beforeEach } from 'node:test';
 import ora from 'ora';
 
 import { BusinessError } from '../../src/errors';
-import actionRunner from '../../src/services/action-runner';
+import actionRunner from '../../src/dialogs/action-runner';
 
 jest.mock('ora');
 const processExit = jest.spyOn(process, 'exit').mockImplementation();
